Fix wrong packages after refactor
