+++ features/location/impl/src/test/kotlin/io/element/android/features/location/impl/send/SendLocationPresenterTest.kt
@@ -23,11 +23,11 @@ import com.google.common.truth.Truth
 import im.vector.app.features.analytics.plan.Composer
 import io.element.android.features.analytics.test.FakeAnalyticsService
 import io.element.android.features.location.api.Location
+import io.element.android.features.location.impl.common.actions.FakeLocationActions
 import io.element.android.features.location.impl.common.permissions.PermissionsEvents
 import io.element.android.features.location.impl.common.permissions.PermissionsPresenter
-import io.element.android.features.location.impl.permissions.PermissionsPresenterFake
+import io.element.android.features.location.impl.common.permissions.PermissionsPresenterFake
 import io.element.android.features.location.impl.common.permissions.PermissionsState
-import io.element.android.features.location.impl.common.actions.FakeLocationActions
 import io.element.android.features.messages.test.MessageComposerContextFake
 import io.element.android.libraries.matrix.api.room.location.AssetType
 import io.element.android.libraries.matrix.test.core.aBuildMeta
