+++ features/location/impl/src/test/kotlin/io/element/android/features/location/impl/show/ShowLocationPresenterTest.kt
@@ -23,7 +23,7 @@ import com.google.common.truth.Truth
 import io.element.android.features.location.api.Location
 import io.element.android.features.location.impl.common.actions.FakeLocationActions
 import io.element.android.features.location.impl.common.permissions.PermissionsPresenter
-import io.element.android.features.location.impl.permissions.PermissionsPresenterFake
+import io.element.android.features.location.impl.common.permissions.PermissionsPresenterFake
 import io.element.android.features.location.impl.common.permissions.PermissionsState
 import io.element.android.libraries.matrix.test.core.aBuildMeta
 import kotlinx.coroutines.delay
@@ -126,5 +126,4 @@ class ShowLocationPresenterTest {
     companion object {
         private const val A_DESCRIPTION = "My happy place"
     }
-
 }
